feature: fixture 를 utils 로 독립, import 수정
